+++ src/mx_bluesky/I24/serial/setup_beamline/pv.py
@@ -448,7 +448,7 @@ me14e_gp97 = "ME14E-MO-IOC-01:GP97"
 me14e_gp98 = "ME14E-MO-IOC-01:GP98"
 me14e_gp99 = "ME14E-MO-IOC-01:GP99"
 me14e_gp100 = "ME14E-MO-IOC-01:GP100"
-me14e_gp101 = "ME14E-MO-IOC-01:GP101"
+me14e_gp101 = "ME14E-MO-IOC-01:GP101"  # Detector in use
 me14e_gp102 = "ME14E-MO-IOC-01:GP102"
 me14e_gp103 = "ME14E-MO-IOC-01:GP103"
 me14e_gp104 = "ME14E-MO-IOC-01:GP104"
