+++ tests/I24/serial/extruder/test_extruder_collect.py
@@ -47,9 +47,7 @@ def test_scrape_parameter_file():
 
 @patch("mx_bluesky.I24.serial.extruder.i24ssx_Extruder_Collect_py3v2.caget")
 @patch("mx_bluesky.I24.serial.extruder.i24ssx_Extruder_Collect_py3v2.caput")
-@patch(
-    "mx_bluesky.I24.serial.extruder.i24ssx_Extruder_Collect_py3v2.sup.get_detector_type"
-)
+@patch("mx_bluesky.I24.serial.extruder.i24ssx_Extruder_Collect_py3v2.get_detector_type")
 @patch("mx_bluesky.I24.serial.extruder.i24ssx_Extruder_Collect_py3v2.logger")
 def test_initialise_extruder(fake_log, fake_det, fake_caput, fake_caget):
     fake_caget.return_value = "/path/to/visit"
